Change LSP config to use nvim 0.11 features
vim.lsp.config() and vim.lsp.enable() is the new API to configure
and start LSPs. No need for the old nvim-lspconfig framework with
its "setup" calls. The plugin is still useful tho as it keeps the
default configs for many different LSPs and those are used by
default if it's installed.